requirements.txt
asgiref==3.6.0
Django==4.2
django-phone-field==1.8.1
sqlparse==0.4.4
